chain: create primitives module.
This contains definitions of primitive types used in other structures
and re-exports of component libraries.
